Contenidos/07_ia_ds_modulo_7/00_contenido/04_semana_4/cv-webapp/backend/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
python-multipart
pytest
